Move PaintStructKeyGenerator to separate file
