+++ src/openrct2/libopenrct2.vcxproj
@@ -326,6 +326,7 @@
     <ClInclude Include="paint\Paint.h" />
     <ClInclude Include="paint\Painter.h" />
     <ClInclude Include="paint\PaintStructDescriptor.h" />
+    <ClInclude Include="paint\PaintStructKeyGenerator.h" />
     <ClInclude Include="paint\Supports.h" />
     <ClInclude Include="paint\tile_element\Paint.Surface.h" />
     <ClInclude Include="paint\tile_element\Paint.TileElement.h" />
@@ -818,6 +819,7 @@
     <ClCompile Include="paint\Painter.cpp" />
     <ClCompile Include="paint\PaintHelpers.cpp" />
     <ClCompile Include="paint\PaintStructDescriptor.cpp" />
+    <ClCompile Include="paint\PaintStructKeyGenerator.cpp" />
     <ClCompile Include="paint\Supports.cpp" />
     <ClCompile Include="paint\tile_element\Paint.Banner.cpp" />
     <ClCompile Include="paint\tile_element\Paint.Entrance.cpp" />
